JAMES-3574 Rework LMTP test
 - Use of a single test class leveraging JUNIT 5 nested classes
 - Common teardown
 - Reuses code upon setup
 - Use of default in memory mailbox test resources
